qrexec: add option to use single socket for stdin+stdout
Instead of two pipes. This is required to use USBIP over qrexec.

To use this feature, pass NULL to stdout_fd. Then single socket will be
created for both stdin and stdout. Open FD to it will be returned in
stdin_fd. Currently it is used only by qrexec-agent.

When stdout_fd!=NULL, old approach will be used (two pipes). This is
currently the case for qrexec-daemon.